[ADD] seeders for database test
