--- backend/database/seeders/20220406075840-Images.js
@@ -1,26 +1,0 @@
-"use strict";
-
-module.exports = {
-    async up(queryInterface, Sequelize) {
-        await queryInterface.bulkInsert(
-            "images",
-            [
-                {
-                    car_id: 1,
-                    file_names: ["imagesblabla.jpg", "ramadan.jpg"]
-                }
-            ],
-            {}
-        );
-    },
-
-    async down(queryInterface, Sequelize) {
-        /**
-         * Add commands to revert seed here.
-         *
-         * Example:
-         * await queryInterface.bulkDelete('People', null, {});
-         */
-        await queryInterface.bulkDelete("images", null, {});
-    }
-};
