--- backend/database/seeders/20220406080733-Cars.js
@@ -1,31 +1,0 @@
-"use strict";
-
-module.exports = {
-    async up(queryInterface, Sequelize) {
-        await queryInterface.bulkInsert(
-            "cars",
-            [
-                {
-                    name: "merco",
-                    price: 150,
-                    brand_id: 1,
-                    color: "rouge",
-                    doors: 5,
-                    boot_size: 143,
-                    type: "sportif",
-                    energy: "Essence",
-                    is_automatic: true,
-                    air_conditioning: true,
-                    is_available: true,
-                    passengers: 4,
-                    description: "belle voiture"
-                }
-            ],
-            {}
-        );
-    },
-
-    async down(queryInterface, Sequelize) {
-        await queryInterface.bulkDelete("cars", null, {});
-    }
-};
