--- backend/database/seeders/20220406080738-CarsBrands.js
@@ -1,20 +1,0 @@
-"use strict";
-
-module.exports = {
-    async up(queryInterface, Sequelize) {
-        await queryInterface.bulkInsert(
-            "cars_brands",
-            [
-                {
-                    brand: "Mercedes",
-                    model: "Class A"
-                }
-            ],
-            {}
-        );
-    },
-
-    async down(queryInterface, Sequelize) {
-        await queryInterface.bulkDelete("cars_brands", null, {});
-    }
-};
