+++ backend/database/seeders/test/20220504095846-test-images.js
@@ -1,0 +1,26 @@
+"use strict";
+
+module.exports = {
+	async up(queryInterface, Sequelize) {
+		await queryInterface.bulkInsert(
+			"images",
+			[
+				{
+					car_id: 1,
+					file_names: ["imagesblabla.jpg", "ramadan.jpg"],
+				},
+			],
+			{}
+		);
+	},
+
+	async down(queryInterface, Sequelize) {
+		/**
+		 * Add commands to revert seed here.
+		 *
+		 * Example:
+		 * await queryInterface.bulkDelete('People', null, {});
+		 */
+		await queryInterface.bulkDelete("images", null, {});
+	},
+};
