+++ backend/database/seeders/test/20220504095917-test-cars-brands.js
@@ -1,0 +1,20 @@
+"use strict";
+
+module.exports = {
+	async up(queryInterface, Sequelize) {
+		await queryInterface.bulkInsert(
+			"cars_brands",
+			[
+				{
+					brand: "Mercedes",
+					model: "Class A",
+				},
+			],
+			{}
+		);
+	},
+
+	async down(queryInterface, Sequelize) {
+		await queryInterface.bulkDelete("cars_brands", null, {});
+	},
+};
